+++ src/main/java/com/comoressoft/mybudget/dto/TotalSummaryDTO.java
@@ -1,33 +1,31 @@
 package com.comoressoft.mybudget.dto;
 
+import java.math.BigDecimal;
+
 public class TotalSummaryDTO {
 
-	private Float income;
-	private Float expense;
-	private Float balance;
-
-	public Float getIncome() {
+	private BigDecimal income;
+	private BigDecimal expense;
+	private BigDecimal balance;
+	public BigDecimal getIncome() {
 		return income;
 	}
-
-	public void setIncome(Float income) {
+	public void setIncome(BigDecimal income) {
 		this.income = income;
 	}
-
-	public Float getExpense() {
+	public BigDecimal getExpense() {
 		return expense;
 	}
-
-	public void setExpense(Float expense) {
+	public void setExpense(BigDecimal expense) {
 		this.expense = expense;
 	}
-
-	public Float getBalance() {
+	public BigDecimal getBalance() {
 		return balance;
 	}
-
-	public void setBalance(Float balance) {
+	public void setBalance(BigDecimal balance) {
 		this.balance = balance;
 	}
+
+	
 
 }
